# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "hello world service"
  version: "1.0.0"
  title: "Hello World API"
  contact:
    email: "you@your-company.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "Post"
  description: "Everything about Posts"
- name: "User"
  description: "Evertyhign about Users"
paths:
  /post:
    x-swagger-router-controller: 'Post'
    get:
      tags:
      - "Post"
      summary: "Fetch posts"
      operationId: "getPost"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "a list of posts to be returned"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Post"
        404:
          description: "Posts not found"
    post:
      tags:
      - "Post"
      summary: "Add a new post"
      operationId: "postPost"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "post"
        required: false
        schema:
          $ref: "#/definitions/NewPost"
      responses:
        200:
          description: "added post"
          schema:
            $ref: "#/definitions/Post"
        400:
          description: "Invalid ID or body supplied"
      security:
      - post_auth:
        - "write:post"
  /post/{id}:
    put:
      tags:
      - "Post"
      summary: "Update a post"
      operationId: "updatePost"
      parameters:
      - name: "id"
        in: "path"
        description: "Post id"
        required: true
        type: "integer"
      - in: "body"
        name: "post"
        required: false
        schema:
          $ref: "#/definitions/NewPost"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Post"
        400:
          description: "Invalid ID or body supplied"
      security:
      - post_auth:
        - "write:post"
    delete:
      tags:
      - "Post"
      summary: "Deletes a post"
      operationId: "deletePost"
      parameters:
      - name: "id"
        in: "path"
        description: "Post id"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid ID supplied"
      security:
      - post_auth:
        - "write:post"
securityDefinitions:
  post_auth:
    type: "oauth2"
    tokenUrl: "http://e302880e.ngrok.io"
    flow: "password"
    scopes:
      read:post: "read post"
      write:post: "wrtie post"
definitions:
  NewPost:
    type: "object"
    properties:
      title:
        type: "string"
      body:
        type: "string"
      userId:
        type: "integer"
        format: "int32"
    example:
      title: "title"
      body: "body"
      userId: 0
  Post:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
        example: 20
      title:
        type: "string"
        example: "Hello World"
      body:
        type: "string"
        example: "Lorem ipsum dolor isit."
      userId:
        type: "integer"
        format: "int32"
        example: 1
    example:
      id: 20
      title: "Hello World"
      body: "Lorem ipsum dolor isit."
      userId: 1